RF | Structure > Created common DeleteItemPage
