+++ src/test/java/model/ManageUsersPage.java
@@ -58,10 +58,10 @@ public class ManageUsersPage extends BasePage {
         return new HomePage(getDriver());
     }
 
-    public DeleteUserPage clickDeleteUser(String name) {
+    public DeleteItemPage clickDeleteUser(String name) {
         getWait(3).until(ExpectedConditions.elementToBeClickable(By.cssSelector("a[href='user/" + name.toLowerCase() + "/delete']"))).click();
 
-        return new DeleteUserPage(getDriver());
+        return new DeleteItemPage(getDriver());
     }
 
     public ManageUsersPage clickUserDropdownMenu(){
